requirements.txt
pymongo>=4.5.0
psycopg2-binary>=2.9.6
schedule>=1.2.0
psutil>=5.9.5
python-dotenv>=1.0.0